+++ src/core/models/iforge-edp-value.mjs
@@ -1,11 +1,13 @@
 import { assertCatalogValue } from "../base/assert-catalog-value.mjs";
-import { valueTypeCatalog } from "../../constants/value-types.mjs";
+import { VALUE_TYPES, valueTypeCatalog } from "../../constants/value-types.mjs";
 import { quantityTypeCatalog } from "../../constants/quantity-types.mjs";
 
 export class IForgeEdpValue {
   constructor({ valueType, quantity, input, internal }) {
     assertCatalogValue(valueType, valueTypeCatalog, "VALUE_TYPES");
-    assertCatalogValue(quantity, quantityTypeCatalog, "QUANTITY_TYPES");
+    if (valueType !== VALUE_TYPES.STRING) {
+      assertCatalogValue(quantity, quantityTypeCatalog, "QUANTITY_TYPES");
+    }
 
     this.valueType = valueType;
     this.quantity = quantity;